renamed aas module to xfsc_aas
